+++ OTAnalytics/plugin_ui/nicegui_gui/pages/add_track_form/add_tracks_form.py
