+++ OTAnalytics/plugin_ui/nicegui_gui/pages/add_track_form/offset_slider_form.py
