+++ OTAnalytics/plugin_ui/nicegui_gui/pages/analysis_form/container.py
